requirements.txt
asgiref==3.2.10
astroid==2.3.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.24.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.11.2
zipp==3.1.0
